XORP/Win32 merge back to mainline.  Bring in Windows specific support routines.   Windows I/O event dispatcher, gai_strerror() stub for MinGW,   non-blocking emulations of read()/write() for Win32 pipes and consoles.  Bring in portable implementations of random(), ether_ntoa(),  ether_aton(), inet_ntop(), inet_pton().  XorpFd file descriptor abstraction.  AsyncFileOperator portability changes.  IoEvent shift to EventLoop, hides SelectorList API.  Portability changes for obtaining system time.  Hack for Linux 2.6 ethernet headers.  Add execute bit to test_run_command.sh.  Make pif and vif indexes 32 bits wide.  Use xorp_osdep headers to deal with platform namespace clashes.
git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6779 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/libxorp/vif.hh
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-// $XORP: xorp/libxorp/vif.hh,v 1.10 2005/01/22 07:05:42 pavlin Exp $
+// $XORP: xorp/libxorp/vif.hh,v 1.11 2005/03/25 02:53:49 pavlin Exp $
 
 #ifndef __LIBXORP_VIF_HH__
 #define __LIBXORP_VIF_HH__
@@ -244,21 +244,21 @@ public:
      * 
      * @return the physical interface index (if applicable).
      */
-    uint16_t pif_index() const { return (_pif_index); }
+    uint32_t pif_index() const { return (_pif_index); }
     
     /**
      * Set the physical interface index.
      * 
      * @param v the value of the physical interface index to set to.
      */
-    void set_pif_index(uint16_t v) { _pif_index = v; }
+    void set_pif_index(uint32_t v) { _pif_index = v; }
     
     /**
      * Various vif_index related values.
      */
     enum {
-	VIF_INDEX_INVALID	= ((uint16_t)~0),
-	VIF_INDEX_MAX		= ((uint16_t)~0)
+	VIF_INDEX_INVALID	= ((uint32_t)~0),
+	VIF_INDEX_MAX		= ((uint32_t)~0)
     };
     
     /**
@@ -266,14 +266,14 @@ public:
      * 
      * @return the virtual interface index.
      */
-    const uint16_t vif_index() const { return (_vif_index); }
+    const uint32_t vif_index() const { return (_vif_index); }
     
     /**
      * Set the virtual interface index.
      * 
      * @param v the value of the virtual interface index to set to.
      */
-    void set_vif_index(uint16_t v) { _vif_index = v; }
+    void set_vif_index(uint32_t v) { _vif_index = v; }
 
     /**
      * Test if this vif is a PIM Register interface.
@@ -501,9 +501,9 @@ public:
 private:
     string	_name;		// The vif name
     string	_ifname;	// The physical interface name (if applicable)
-    uint16_t	_pif_index;	// Physical interface index (if applicable),
+    uint32_t	_pif_index;	// Physical interface index (if applicable),
 				// or 0 if invalid.
-    uint16_t	_vif_index;	// Virtual interface index
+    uint32_t	_vif_index;	// Virtual interface index
     bool	_is_pim_register;	// PIM Register vif
     bool	_is_p2p;		// Point-to-point interface
     bool	_is_loopback;		// Loopback interface
